Unify icon url generation at central place
