+++ adhocracy4/maps/templatetags/maps_tags.py
@@ -2,9 +2,10 @@ import json
 
 from django import template
 from django.conf import settings
-from django.contrib.staticfiles.templatetags.staticfiles import static
 from django.utils.html import format_html
 from easy_thumbnails.files import get_thumbnailer
+
+from adhocracy4.categories import get_category_pin_url
 
 register = template.Library()
 
@@ -32,8 +33,7 @@ def get_points(items):
             negative_rating_count = item.negative_rating_count
 
         if hasattr(item, 'category') and getattr(item.category, 'icon', None):
-            category_icon = static('category_icons/pins/{}_pin.svg'.
-                                   format(item.category.icon))
+            category_icon = get_category_pin_url(item.category.icon)
         else:
             category_icon = ''
